Added basically aliases to the bot
* Aliases need to be set via mapcfg by using game_alias,
this is a preparation for later when we include many other modes etc,
the alias will be recorded everywhere so we can easily split stats but
using only one database.

* The alias we are using currently is basically simple and just define
the maptype.
We nowe think, about a database created like that: dota - 1, lod - 2,
later we will add other aliases and support more statstypes like w3mmd.