update
* added reserve slot command for a specific command
* added rcon command to reserve a slot

========================================================
We have seen several pages without any kind of copyrights.
We have started to discuss possible results out of it. It could be
possible that we will switch to a closed source project.

This project is made from us and we are the contributers if you remove
copyrights its like tellings its your product. Be fair, respect our work
and keep copyrights.

We will neither help anyone who had removed once the copyright.